# fix: hash chunks with djb2