--- a/.changeset/tough-timers-admire.md
+++ b/.changeset/tough-timers-admire.md
@@ -1,0 +1,5 @@
+---
+"vinxi": patch
+---
+
+fix: hash chunks with djb2