--- a/packages/vinxi/lib/chunks.test.js
+++ b/packages/vinxi/lib/chunks.test.js
@@ -1,0 +1,9 @@
+import { describe, expect, it } from "vitest";
+
+import { hash } from "./chunks";
+
+describe("hash", () => {
+	it("should create different hashes for abc and cba", () => {
+		expect(hash("abc")).not.toEqual(hash("cba"));
+	});
+});